backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
django==5.1.3
django-environ==0.11.2
django-filter==25.1
django-types==0.19.1
djangorestframework==3.15.2
idna==3.10
iniconfig==2.0.0
jinja2==3.1.6
langcodes==3.5.0
language-data==1.3.0
marisa-trie==1.2.1
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
numpy==2.2.3
packaging==24.2
pdfminer-six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pluggy==1.5.0
preshed==3.0.9
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pypdfium2==4.30.1
pytest==8.3.3
requests==2.32.3
rich==13.9.4
ruff==0.6.9
setuptools==76.0.0
shellingham==1.5.4
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.4
tqdm==4.67.1
typer==0.15.2
types-psycopg2==2.9.21.20250121
typing-extensions==4.12.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
